requirements.txt
absl-py==1.4.0
astroid==2.14.2
astunparse==1.6.3
cachetools==5.3.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cmake==3.25.0
colorama==0.4.6
conda==4.3.16
dbus-python==1.2.16
dill==0.3.6
Flask==2.2.3
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
idna==2.8
importlib-metadata==6.0.0
imutils==0.5.4
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
lazy-object-proxy==1.9.0
libclang==15.0.6.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.24.2
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
platformdirs==3.0.0
playsound==1.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.10.0
pycosat==0.6.3
PyGObject==3.36.0
pylint==2.16.2
PyMySQL==1.0.3
python-apt==2.0.1
python-barcode==0.14.0
python-dotenv==1.0.0
pyttsx3==2.90
regex==2022.10.31
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
spark-parser==1.8.9
SQLAlchemy==2.0.6
unattended-upgrades==0.1
uncompyle6==3.9.0
urllib3==1.25.8
Werkzeug==2.2.3
wrapt==1.14.1
xdis==6.0.5
zipp==3.14.0
